AppWidgetIntent issue solved
When I've created the AppWidget I had a problem about intents. Android reuses intents, so I created a new intent, making sure to put an unique ID (as AppWidgetId) in it, otherwise the same intent and WidgetId would be used as trigger for all instances. Now every widget open the right ChatActivity as selected ib the WidgetConfigurationActivity.